+++ app/helpers/offender_jurisdictions_helper.rb
@@ -1,0 +1,6 @@
+module OffenderJurisdictionsHelper
+  def search_params
+    @_search_params ||=
+      params.fetch(:offender, {}).slice(:first_name, :last_name)
+  end
+end
