+++ spec/controllers/offenders_controller_spec.rb
@@ -40,49 +40,6 @@ DCJ_OFFENDER = {
 RSpec.describe OffendersController do
   render_views
 
-  describe 'GET /index' do
-    let(:params) { {} }
-
-    subject { get :index, params: params }
-
-    it 'renders successfully' do
-      subject
-      expect(response).to be_success
-    end
-
-    describe 'with search by SID parameters' do
-      let(:params) { { offender: { sid: '1234' } } }
-
-      it 'redirects to the offender show page' do
-        subject
-        expect(response).to redirect_to(offender_path(:oregon, params[:offender][:sid]))
-      end
-    end
-
-    describe 'with search by name' do
-      let(:params) { { offender: { first_name: 'Tom', last_name: 'Dooner' } } }
-      let(:results) { [{ sid: 123456, jurisdiction: :oregon, first: 'Tom', last: 'Dooner' }] }
-
-      before do
-        allow(OffenderScraper).to receive(:search_by_name).and_return(results)
-      end
-
-      it 'renders the results' do
-        subject
-        expect(response.body).to include(results[0][:sid].to_s)
-      end
-
-      describe 'when there are no results' do
-        let(:results) { [] }
-
-        it 'gives an error' do
-          subject
-          expect(response.body).to include("We couldn't find that offender.")
-        end
-      end
-    end
-  end
-
   describe 'GET /show' do
     before do
       allow(OffenderScraper).to receive(:offender_details)
